Simplified code dramatically. Should now work for most normal pages, if the currency is implemented. Version 0.2.0
It turns out that there is a table surrounding the text I need, so I can safely use regex on the whole table for the bid amount. My testing suggests that there is no problem using getElementsByName in Firefox (at least my version), so I've switched to that for the bid box. Between these, it should mean I no longer need to deal with all the changing ids. Great!